pyproject.toml
[project]
name = "aifoundry-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-search-documents>=11.5.0",
    "azure-identity>=1.17.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "openai>=1.84.0",
    "azure-ai-projects>=1.0.0b11",
    "azure-monitor-opentelemetry>=1.6.10",
    "opentelemetry-sdk>=1.31.1",
    "opentelemetry-instrumentation-openai>=0.32.0",
    "opentelemetry-instrumentation-requests>=0.49.0",
    "opentelemetry-instrumentation>=0.49.0",
    "azure-ai-evaluation>=1.8.0",
]
